1. Turn endpoint into a pipeline
The Endpoint is the entry point for a request in the Phoenix pipeline.
You'll start to see this theme: everything in Phoenix is a pipeline. The
endpoint, the router, the controller. The endpoint is pluggable so that
other transformations may be added ahead of routing requests. Think of
this much like Rack middelware.