Move Feenix Params module out of Controller namespace
When I was planning this originally, I thought that Phoenix controllers
handled fetching parameters. It actually seems like this is done in part
by plugging Plug.Parsers in the _app_ Endpoint and in other part during
routing. To clarify that this implementation really is independent of
the controller, I'm moving it out of that namespace.